src/test_postgres_redis/requirements.txt
psycopg2-binary
pydantic
python-dotenv